requirements.txt
langchain==0.1.1
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
huggingface-hub==0.17.3
streamlit==1.28.2
langchain-community==0.0.13
python-dotenv==1.0.0